# MEDRISE-MEDICAL-CENTRE-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedRise Medical Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff accounts
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a new staff account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "400":
          description: Invalid input
        "409":
          description: Username already exists
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "404":
          description: Not found
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "400":
          description: Cannot delete own account
        "404":
          description: Not found
...
components:
  schemas:
...
    StaffMember:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
        title:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
    StaffInput:
      type: object
      required: [username, password, name, role]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        name:
          type: string
          minLength: 2
        role:
          type: string
          enum: [admin, doctor, nurse, midwife, receptionist, staff]
        title:
          type: string
        phone:
          type: string
        email:
          type: string
    StaffUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, doctor, nurse, midwife, receptionist, staff]
        title:
          type: string
        phone:
          type: string
        email:
          type: string
...